Add files via upload
Write an if statement that checks if a number is even or odd. 
Use nested if statements to classify a number as negative, positive, or zero. 
Use the conditional (ternary) operator ‘?’ to rewrite a simple if…else statement. 
Check the validity of a variable using the ? operator. 
Use the conditional operator to assign a value to a variable based on a 
condition.